pyproject.toml
[tool.poetry]
name = "filmstore"
version = "0.1.0"
description = "A film api built with Django"
authors = ["Louis Agyapong <agyapong.louis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-extensions = "^3.1.5"
python-decouple = "^3.6"
ipython = "^8.3.0"
dj-database-url = "^0.5.0"
black = "^22.3.0"
mypy = "^0.960"
isort = "^5.10.1"
djangorestframework = "^3.13.1"
django-taggit = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
